% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{agony}
\alias{agony}
\title{agony}
\usage{
agony(inname, outname)
}
\arguments{
\item{inname}{Input agony file.}

\item{outname}{Output agony file.}
}
\description{
Perform agony computation. The software for agony computation (exact algorithm) adopted here was developed by Professor Nikolaj Tatti and colleagues.
For a detailed description, please refer to: Tatti, Nikolaj. "Tiers for peers: a practical algorithm for discovering hierarchy in weighted networks." Data mining and knowledge discovery 31.3 (2017): 702-738.
}
