% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agonyArcs}
\alias{agonyArcs}
\title{Example dataset to test the agony function.}
\format{
Example dataset to test the agony function.
}
\source{
Example dataset to test the agony function.
}
\usage{
data(agonyArcs)
}
\value{
Example dataset to test the agony function.
}
\description{
Example dataset to test the agony function.
}
