% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aml_example}
\alias{aml_example}
\title{Example dataset for 366 patients with Acute Myeloid Leukaemia.}
\format{
List of two elements, i.e., survival data and evolutionary steps extracted by ASCETIC.
}
\source{
Tyner, Jeffrey W., et al. "Functional genomic landscape of acute myeloid leukaemia." Nature 562.7728 (2018): 526-531.
}
\usage{
data(aml_example)
}
\value{
List of two elements, i.e., survival data and evolutionary steps extracted by ASCETIC.
}
\description{
366 patients with Acute Myeloid Leukaemia obtained from Tyner, Jeffrey W., et al. 2018.
}
