% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascetic.R
\name{asceticPhylogeniesAssessment}
\alias{asceticPhylogeniesAssessment}
\title{asceticPhylogeniesAssessment}
\usage{
asceticPhylogeniesAssessment(inference, niterations = 100, nsampling = 100)
}
\arguments{
\item{inference}{Model inferred with ASCETIC using either the function asceticPhylogeniesBootstrap or the function asceticPhylogenies.}

\item{niterations}{Number of cross-validation iterations to be performed for a robust assessment of ASCETIC model.
Higher values lead to improved estimates, but require higher computational burden; default value is 100.}

\item{nsampling}{Number of re-sampling to be performed for a robust estimation of the agony ranking.
Higher values lead to improved estimates, but require higher computational burden; default value is 100.}
}
\value{
A list of 3 elements for which the estimate by cross-validation is performed: 
                              1) rankingEstimate, ranking among mutations estimated by agony.
                                 Lower rankings correspond to early mutations. This is returned only if nsampling > 0.
                              2) poset, partially order set among mutations estimated by ASCETIC from the agony ranking.
                              3) inference, inferred ASCETIC evolutionary model for each selected regularization.
}
\description{
Perform the assessment via cross-validation of the model inferred by the ASCETIC framework on 
multiple samples (using mutational trees as inputs) datasets.
}
\examples{
set.seed(12345)
data(datasetExamplePhylogenies)
data(modelsPhylogenies)
resExamplePhylogeniesDatasetBootstrap <- asceticPhylogeniesBootstrap(
                                                    dataset = datasetExamplePhylogenies,
                                                    models = modelsPhylogenies,
                                                    nsampling = 3,
                                                    regularization = c("aic","bic"),
                                                    command = "hc",
                                                    restarts = 0 )
resExamplePhylogeniesAssessment <- asceticPhylogeniesAssessment(
                                          inference = resExamplePhylogeniesDatasetBootstrap,
                                          niterations = 3,
                                          nsampling = 3)

}
