% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascetic.R
\name{asceticPhylogeniesBootstrap}
\alias{asceticPhylogeniesBootstrap}
\title{asceticPhylogeniesBootstrap}
\usage{
asceticPhylogeniesBootstrap(
  dataset,
  models,
  nsampling = 100,
  regularization = c("aic", "bic"),
  command = "hc",
  restarts = 10
)
}
\arguments{
\item{dataset}{Binary matrix where rows are samples and columns are mutations.
Each cell of the matrix is 1 if the related mutation was observed in the sample; 0 otherwise.}

\item{models}{Phylogenetic models estimated for each sample included in dataset.}

\item{nsampling}{Number of re-sampling to be performed for a robust estimation of the agony ranking.
Higher values lead to improved estimates, but require higher computational burden; default value is 100.}

\item{regularization}{Regularization to be used for the maximum likelihood estimation.
Possible values are aic for the Akaike information criterion and bic for the Bayesian information criterion.
For the complete list of options, we refer to the manual of the bnlearn package.}

\item{command}{Optimization technique to be used for maximum likelihood estimation.
Valid values are either hc for Hill Climbing or tabu for Tabu Search.}

\item{restarts}{Number of restarts to be performed during the maximum likelihood estimation when Hill Climbing optimization technique is used.
Higher values lead to improved estimates, but require higher computational burden; default value is 10.
This parameter is ignored if tabu search is selected.}
}
\value{
A list of 5 elements: 1) dataset, input dataset.
                              2) models, input phylogenetic models
                              3) ranking_estimate, ranking among mutations estimated by agony.
                                 Lower rankings correspond to early mutations.
                              4) poset, partially order set among mutations estimated by ASCETIC from the agony ranking.
                              5) inference, inferred ASCETIC evolutionary model for each selected regularization.
}
\description{
Perform the ASCETIC inference framework on multiple samples (using phylogenies as inputs) datasets 
with bootstrap for a robust estimation of the agony ranking.
}
\examples{
set.seed(12345)
data(datasetExamplePhylogenies)
data(modelsPhylogenies)
resExamplePhylogeniesDatasetBootstrap <- asceticPhylogeniesBootstrap(
                                                    dataset = datasetExamplePhylogenies,
                                                    models = modelsPhylogenies,
                                                    nsampling = 5,
                                                    regularization = "aic",
                                                    command = "hc",
                                                    restarts = 0 )

}
