% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build.agony.input.multiple.samples}
\alias{build.agony.input.multiple.samples}
\title{build.agony.input.multiple.samples}
\usage{
build.agony.input.multiple.samples(models, events)
}
\arguments{
\item{models}{Phylogenetic models estimated for each sample included in dataset.}

\item{events}{Driver genes to be considered in the inference.}
}
\description{
Compute a set of time orderings among each event, given a set of models inferred from multiple samples
}
