% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build.ranking.adj.matrix}
\alias{build.ranking.adj.matrix}
\title{build.ranking.adj.matrix}
\usage{
build.ranking.adj.matrix(agony_ranking, num_events)
}
\arguments{
\item{agony_ranking}{Agony-based rankings.}

\item{num_events}{Number of driver genes considered for the inference.}
}
\description{
Compute the poset based on the best agony ranking
}
