% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute.agony.poset}
\alias{compute.agony.poset}
\title{compute.agony.poset}
\usage{
compute.agony.poset(
  agony_arcs,
  num_events,
  agony_files = paste0(getwd(), "/agony_files")
)
}
\arguments{
\item{agony_arcs}{Agony-based rankings.}

\item{num_events}{Number of driver genes considered for the inference.}

\item{agony_files}{Location of agony files.}
}
\description{
Estimate an optimal agony ranking poset given a set of temporal observations.
}
