% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataExamples}
\alias{dataExamples}
\title{List data structure to run examples}
\format{
List data structure to run examples.
}
\source{
List data structure to run examples.
}
\usage{
data(dataExamples)
}
\value{
List data structure to run examples.
}
\description{
List data structure to run examples.
}
