% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset_example_single_samples}
\alias{dataset_example_single_samples}
\title{Binary matrix for 10 hepatocellular carcinoma sequencing samples.}
\format{
Binary matrix where rows are samples and columns are mutations.
}
\source{
Nguyen, Bastien, et al. "Genomic characterization of metastatic patterns from prospective clinical sequencing of 25,000 patients." Cell 185.3 (2022): 563-575.
}
\usage{
data(dataset_example_single_samples)
}
\value{
Binary matrix where rows are samples and columns are mutations.
}
\description{
Binary matrix for 10 hepatocellular carcinoma sequencing samples from Nguyen, Bastien, et al., 2022.
}
