% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasetExamplePhylogenies}
\alias{datasetExamplePhylogenies}
\title{Binary matrix for 10 myeloid malignancies samples sequenced with Tapestri.}
\format{
Binary matrix where rows are samples and columns are mutations.
}
\source{
Miles, Linde A., et al. "Single-cell mutation analysis of clonal evolution in myeloid malignancies." Nature 587.7834 (2020): 477-482.
}
\usage{
data(datasetExamplePhylogenies)
}
\value{
Binary matrix where rows are samples and columns are mutations.
}
\description{
Binary matrix for 10 myeloid malignancies samples sequenced with Tapestri from Miles, Linde A., et al., 2020.
}
