% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.applyPr}
\alias{.applyPr}
\title{.applyPr}
\usage{
.applyPr(poset, prModel, prNull)
}
\arguments{
\item{poset}{Partially order set to be considered during the inference.}

\item{pr_model}{ASCETIC statistical model based on Suppes' theory of probabilistic causation.}

\item{pr_null}{ASCETIC null model based on Suppes' theory of probabilistic causation.}
}
\description{
Enforce probability raising constraints.
}
\examples{
data(dataExamples)
dataExample <- dataExamples[["apply.pr"]]
res <- .applyPr(poset=dataExample$poset,pr_model=dataExample$pr_model,pr_null=dataExample$pr_null)

}
