% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.asCategoricalDataset}
\alias{.asCategoricalDataset}
\title{as.categorical.dataset}
\usage{
.asCategoricalDataset(dataset)
}
\arguments{
\item{dataset}{Binary matrix where rows are samples and columns are mutations.}
}
\description{
Create a categorical data structure.
}
\examples{
data(dataExamples)
dataExample <- dataExamples[["as.categorical.dataset"]]
res <- as.categorical.dataset(dataExample)

}
