% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.buildAgonyInput}
\alias{.buildAgonyInput}
\title{.buildAgonyInput}
\usage{
.buildAgonyInput(agonyOrders)
}
\arguments{
\item{agonyOrders}{Agony-based rankings.}
}
\description{
Create the input file for computing a best agony ranking given a total ordering.
}
\examples{
data(dataExamples)
dataExample <- dataExamples[["build.agony.input"]]
res <- .buildAgonyInput(dataExample)

}
