% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.buildAgonyInputMultipleSamples}
\alias{.buildAgonyInputMultipleSamples}
\title{.buildAgonyInputMultipleSamples}
\usage{
.buildAgonyInputMultipleSamples(models, events)
}
\arguments{
\item{models}{Phylogenetic models estimated for each sample included in dataset.}

\item{events}{Driver genes to be considered in the inference.}
}
\description{
Compute a set of time orderings among each event, given a set of models inferred from multiple samples.
}
\examples{
data(dataExamples)
dataExample <- dataExamples[["build.agony.input.multiple.samples"]]
res <- .buildAgonyInputMultipleSamples(dataExample$models,dataExample$events)

}
