% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.buildRankingAdjMatrix}
\alias{.buildRankingAdjMatrix}
\title{.buildRankingAdjMatrix}
\usage{
.buildRankingAdjMatrix(agonyRanking, numEvents)
}
\arguments{
\item{agonyRanking}{Agony-based rankings.}

\item{num_events}{Number of driver genes considered for the inference.}
}
\description{
Compute the poset based on the best agony ranking.
}
\examples{
data(dataExamples)
dataExample <- dataExamples[["build.ranking.adj.matrix"]]
res <- .buildRankingAdjMatrix(agonyRanking=dataExample$agony_ranking,num_events=dataExample$num_events)

}
