% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.computeAgonyPoset}
\alias{.computeAgonyPoset}
\title{.computeAgonyPoset}
\usage{
.computeAgonyPoset(agonyArcs, numEvents)
}
\arguments{
\item{agonyArcs}{Agony-based rankings.}

\item{numEvents}{Number of driver genes considered for the inference.}
}
\description{
Estimate an optimal agony ranking poset given a set of temporal observations.
}
