% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.estimateAgonyPosetSingleSamples}
\alias{.estimateAgonyPosetSingleSamples}
\title{.estimateAgonyPosetSingleSamples}
\usage{
.estimateAgonyPosetSingleSamples(ccfDataset)
}
\arguments{
\item{ccfDataset}{Matrix where rows are samples and columns are mutations.}
}
\description{
Estimate the poset based on a best agony ranking from a dataset of cancer cell fractions.
}
