% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.estimatePrModelMultipleSamples}
\alias{.estimatePrModelMultipleSamples}
\title{.estimatePrModelMultipleSamples}
\usage{
.estimatePrModelMultipleSamples(models, events)
}
\arguments{
\item{models}{Phylogenetic models estimated for each sample included in dataset.}

\item{events}{Driver genes to be considered in the inference.}
}
\description{
Estimate the model for the probability raising, i.e., j --> i, P(t_j<t_i|i,j), from a inferences on multiple patients.
}
\examples{
data(dataExamples)
dataExample <- dataExamples[["estimate.pr.model.multiple.samples"]]
res <- .estimatePrModelMultipleSamples(dataExample$models,dataExample$events)

}
