% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.estimatePrModelSingleSamples}
\alias{.estimatePrModelSingleSamples}
\title{.estimatePrModelSingleSamples}
\usage{
.estimatePrModelSingleSamples(ccfDataset)
}
\arguments{
\item{ccfDataset}{Matrix where rows are samples and columns are mutations.}
}
\description{
Estimate the model for the probability raising, i.e., j --> i, P(t_j<t_i|i,j), from a dataset of cancer cell fractions.
}
\examples{
data(dataExamples)
dataExample <- dataExamples[["estimate.pr.model.single.samples"]]
res <- .estimatePrModelSingleSamples(ccf.dataset=dataExample)

}
