% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.estimatePrNull}
\alias{.estimatePrNull}
\title{.estimatePrNull}
\usage{
.estimatePrNull(dataset)
}
\arguments{
\item{dataset}{Binary matrix where rows are samples and columns are mutations.}
}
\description{
Estimate the null model for the probability raising, i.e., j --> i, P(i,not j).
}
\examples{
data(dataExamples)
dataExample <- dataExamples[["estimate.pr.null"]]
res <- .estimatePrNull(dataExample)

}
