% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.estimateProbs}
\alias{.estimateProbs}
\title{.estimateProbs}
\usage{
.estimateProbs(dataset)
}
\arguments{
\item{dataset}{Binary matrix where rows are samples and columns are mutations.}
}
\description{
Estimate marginal and joint probabilities from the dataset.
}
\examples{
data(dataExamples)
dataExample <- dataExamples[["estimate.probs"]]
res <- .estimateProbs(dataExample)

}
