% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.inferOrderInCcfDataset}
\alias{.inferOrderInCcfDataset}
\title{.inferOrderInCcfDataset}
\usage{
.inferOrderInCcfDataset(ccfDataset)
}
\arguments{
\item{ccfDataset}{Matrix where rows are samples and columns are mutations.}
}
\description{
Estimate a total ordering for a dataset given the observed cancer cell fractions in each sample.
}
\examples{
data(dataExamples)
dataExample <- dataExamples[["infer.order.in.ccf.dataset"]]
res <- .inferOrderInCcfDataset(dataExample)

}
