% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.inferTotalOrder}
\alias{.inferTotalOrder}
\title{.inferTotalOrder}
\usage{
.inferTotalOrder(ccfEntry)
}
\arguments{
\item{ccfEntry}{Cancer cell fractions for each sample.}
}
\description{
Estimate a total ordering for a sample given the observed cancer cell fractions.
}
\examples{
data(dataExamples)
dataExample <- dataExamples[["infer.total.order"]]
res <- .inferTotalOrder(dataExample)

}
