% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate.agony.poset.multiple.samples}
\alias{estimate.agony.poset.multiple.samples}
\title{estimate.agony.poset.multiple.samples}
\usage{
estimate.agony.poset.multiple.samples(models, events)
}
\arguments{
\item{models}{Phylogenetic models estimated for each sample included in dataset.}

\item{events}{Driver genes to be considered in the inference.}
}
\description{
Estimate the poset based on a best agony ranking from a set of models inferred from multiple samples
}
