% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate.agony.poset.single.samples}
\alias{estimate.agony.poset.single.samples}
\title{estimate.agony.poset.single.samples}
\usage{
estimate.agony.poset.single.samples(ccf.dataset)
}
\arguments{
\item{ccf.dataset}{Matrix where rows are samples and columns are mutations.}
}
\description{
Estimate the poset based on a best agony ranking from a dataset of cancer cell fractions.
}
