% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate.pr.model.single.samples}
\alias{estimate.pr.model.single.samples}
\title{estimate.pr.model.single.samples}
\usage{
estimate.pr.model.single.samples(ccf.dataset)
}
\arguments{
\item{ccf.dataset}{Matrix where rows are samples and columns are mutations.}
}
\description{
Estimate the model for the probability raising, i.e., j --> i, P(t_j<t_i|i,j), from a dataset of cancer cell fractions.
}
