% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate.pr.null}
\alias{estimate.pr.null}
\title{estimate.pr.null}
\usage{
estimate.pr.null(dataset)
}
\arguments{
\item{dataset}{Binary matrix where rows are samples and columns are mutations.}
}
\description{
Estimate the null model for the probability raising, i.e., j --> i, P(i,not j)
}
