% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evosigs.R
\name{evosigs}
\alias{evosigs}
\title{evosigs}
\usage{
evosigs(survival_data, evolutionary_steps)
}
\arguments{
\item{survival_data}{Survival data. Notice that no missing values are allowed.}

\item{evolutionary_steps}{Evolutionary steps per patient as inferred by ASCETIC.}
}
\value{
A list of 3 elements: 1) survival_analysis, results from survival analysis.
                              2) evolutionary_steps, evolutionary steps significantly associated with survival data.
                              3) clusters_prevalence, prevalence per cluster of the selected evolutionary steps.
}
\description{
Select evolutionary steps from ASCETIC analysis which are significantly associated with survival data.
}
\examples{
set.seed(12345)
data(aml_example)
res_example_evosigs = evosigs( 
                              survival_data = aml_example$survival_data, 
                              evolutionary_steps = aml_example$evolutionary_steps )

}
