% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infer.order.in.ccf.dataset}
\alias{infer.order.in.ccf.dataset}
\title{infer.order.in.ccf.dataset}
\usage{
infer.order.in.ccf.dataset(ccf.dataset)
}
\arguments{
\item{ccf.dataset}{Matrix where rows are samples and columns are mutations.}
}
\description{
Estimate a total ordering for a dataset given the observed cancer cell fractions in each sample
}
