% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infer.total.order}
\alias{infer.total.order}
\title{infer.total.order}
\usage{
infer.total.order(ccf.entry)
}
\arguments{
\item{ccf.entry}{Cancer cell fractions for each sample.}
}
\description{
Estimate a total ordering for a sample given the observed cancer cell fractions
}
