% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{models_phylogenies}
\alias{models_phylogenies}
\title{Phylogenetic models estimated for each sample included in dataset.}
\format{
R list of 10 elements providing the phylogenetic models estimated for the samples in the dataset.
}
\source{
Miles, Linde A., et al. "Single-cell mutation analysis of clonal evolution in myeloid malignancies." Nature 587.7834 (2020): 477-482.
}
\usage{
data(models_phylogenies)
}
\value{
R list of 10 elements providing the phylogenetic models estimated for the samples in the dataset.
}
\description{
Phylogenetic models estimated for each sample included in dataset from Miles, Linde A., et al., 2020.
}
