% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perform.likelihood.fit}
\alias{perform.likelihood.fit}
\title{perform.likelihood.fit}
\usage{
perform.likelihood.fit(dataset, poset, regularization, command, restarts)
}
\arguments{
\item{dataset}{Binary matrix where rows are samples and columns are mutations.}

\item{poset}{Partially order set to be considered during the inference.}

\item{regularization}{Regularization to be used for the maximum likelihood estimation.}

\item{command}{Optimization technique to be used for maximum likelihood estimation.}

\item{restarts}{Number of restarts to be performed during the maximum likelihood estimation when Hill Climbing optimization technique is used.}
}
\description{
Perform structure learning by maximum likelihood given a poset.
}
\examples{
data(data_examples)
data_example = data_examples[["perform.likelihood.fit"]]
res = perform.likelihood.fit( dataset = data_example$dataset, 
                              poset = data_example$poset, 
                              regularization = "aic", 
                              command = "hc", 
                              restarts = 0)

}
