% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_example_phylogenies_dataset_bootstrap}
\alias{res_example_phylogenies_dataset_bootstrap}
\title{Example of results obtained by running the function perform.ascetic.phylogenies.dataset.bootstrap on dataset_example_phylogenies}
\format{
Results by perform.ascetic.phylogenies.dataset.bootstrap on dataset_example_phylogenies.
}
\usage{
data(res_example_phylogenies_dataset_bootstrap)
}
\value{
Results by perform.ascetic.phylogenies.dataset.bootstrap on dataset_example_phylogenies.
}
\description{
Example of results obtained by running the function perform.ascetic.phylogenies.dataset.bootstrap on dataset_example_phylogenies from Miles, Linde A., et al., 2020.
}
