% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_example_single_samples_resampling}
\alias{res_example_single_samples_resampling}
\title{Example of results obtained by running the function perform.ascetic.ccf.dataset.resampling on dataset_example_single_samples.}
\format{
Results obtained by running the function perform.ascetic.ccf.dataset.resampling on dataset_example_single_samples.
}
\usage{
data(res_example_single_samples_resampling)
}
\value{
Results obtained by running the function perform.ascetic.ccf.dataset.resampling on dataset_example_single_samples.
}
\description{
Example of results obtained by running the function perform.ascetic.ccf.dataset.resampling on dataset_example_single_samples from Nguyen, Bastien, et al., 2022.
}
