% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resExamplePhylogeniesDatasetBootstrap}
\alias{resExamplePhylogeniesDatasetBootstrap}
\title{Example of results obtained by running the function perform.ascetic.phylogenies.dataset.bootstrap on datasetExamplePhylogenies.}
\format{
Results by perform.ascetic.phylogenies.dataset.bootstrap on datasetExamplePhylogenies.
}
\usage{
data(resExamplePhylogeniesDatasetBootstrap)
}
\value{
Results by perform.ascetic.phylogenies.dataset.bootstrap on datasetExamplePhylogenies.
}
\description{
Example of results obtained by running the function perform.ascetic.phylogenies.dataset.bootstrap on datasetExamplePhylogenies from Miles, Linde A., et al., 2020.
}
