% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resExampleSingleSamplesResampling}
\alias{resExampleSingleSamplesResampling}
\title{Example of results obtained by running the function perform.ascetic.ccf.dataset.resampling on datasetExampleSingleSamples.}
\format{
Results obtained by running the function perform.ascetic.ccf.dataset.resampling on datasetExampleSingleSamples.
}
\usage{
data(resExampleSingleSamplesResampling)
}
\value{
Results obtained by running the function perform.ascetic.ccf.dataset.resampling on datasetExampleSingleSamples.
}
\description{
Example of results by the function perform.ascetic.ccf.dataset.resampling on datasetExampleSingleSamples from Nguyen, Bastien, et al., 2022.
}
