% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vaf_dataset_example_single_samples}
\alias{vaf_dataset_example_single_samples}
\title{R data.frame for 10 hepatocellular carcinoma sequencing samples.}
\format{
R data.frame with 8 columns: 1) SAMPLE_ID; 2) GENE_ID; 3) REF_COUNT; 4) ALT_COUNT; 5) COPY_NUMBER; 6) NORMAL_PLOIDY; 7) VAF_ESTIMATE; 8) CCF_ESTIMATE.
}
\source{
Nguyen, Bastien, et al. "Genomic characterization of metastatic patterns from prospective clinical sequencing of 25,000 patients." Cell 185.3 (2022): 563-575.
}
\usage{
data(vaf_dataset_example_single_samples)
}
\value{
R data.frame with 8 columns: 1) SAMPLE_ID; 2) GENE_ID; 3) REF_COUNT; 4) ALT_COUNT; 5) COPY_NUMBER; 6) NORMAL_PLOIDY; 7) VAF_ESTIMATE; 8) CCF_ESTIMATE.
}
\description{
R data.frame for 10 hepatocellular carcinoma sequencing samples from Nguyen, Bastien, et al., 2022.
}
