% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply.pr}
\alias{apply.pr}
\title{apply.pr}
\usage{
apply.pr(poset, pr_model, pr_null)
}
\arguments{
\item{poset}{Partially order set to be considered during the inference.}

\item{pr_model}{ASCETIC statistical model based on Suppes' theory of probabilistic causation.}

\item{pr_null}{ASCETIC null model based on Suppes' theory of probabilistic causation.}
}
\description{
Enforce probability raising constraints.
}
\examples{
data(data_examples)
data_example = data_examples[["apply.pr"]]
res = apply.pr(poset=data_example$poset,pr_model=data_example$pr_model,pr_null=data_example$pr_null)

}
