% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.categorical.dataset}
\alias{as.categorical.dataset}
\title{as.categorical.dataset}
\usage{
as.categorical.dataset(dataset)
}
\arguments{
\item{dataset}{Binary matrix where rows are samples and columns are mutations.}
}
\description{
Create a categorical data structure
}
