% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build.agony.input}
\alias{build.agony.input}
\title{build.agony.input}
\usage{
build.agony.input(agony_orders)
}
\arguments{
\item{agony_orders}{Agony-based rankings.}
}
\description{
Create the input file for computing a best agony ranking given a total ordering.
}
\examples{
data(data_examples)
data_example = data_examples[["build.agony.input"]]
res = build.agony.input(data_example)

}
