% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_examples}
\alias{data_examples}
\title{list data structure to run examples}
\format{
list data structure to run examples
}
\source{
list data structure to run examples.
}
\usage{
data(data_examples)
}
\value{
list data structure to run examples
}
\description{
list data structure to run examples.
}
