% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate.probs}
\alias{estimate.probs}
\title{estimate.probs}
\usage{
estimate.probs(dataset)
}
\arguments{
\item{dataset}{Binary matrix where rows are samples and columns are mutations.}
}
\description{
Estimate marginal and joint probabilities from the dataset
}
