% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evosigs.R
\name{evoSigs}
\alias{evoSigs}
\title{evoSigs}
\usage{
evoSigs(survivalData, evolutionarySteps)
}
\arguments{
\item{survivalData}{Survival data. Notice that no missing values are allowed.}

\item{evolutionarySteps}{Evolutionary steps per patient as inferred by ASCETIC.}
}
\value{
A list of 3 elements: 1) survivalAnalysis, results from survival analysis.
                              2) evolutionarySteps, evolutionary steps significantly associated with survival data.
                              3) clustersPrevalence, prevalence per cluster of the selected evolutionary steps.
}
\description{
Select evolutionary steps from ASCETIC analysis which are significantly associated with survival data.
}
\examples{
set.seed(12345)
data(amlExample)
resExampleEvosigs = evoSigs( survivalData = amlExample$survival_data, 
                             evolutionarySteps = amlExample$evolutionary_steps )

}
